Changes by tzach

+++ src/main/java/com/att/tdp/bisbis10/controller/DishController.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Controller
@@ -49,10 +50,50 @@ public class DishController {
     @DeleteMapping("/restaurants/{id}/dishes/{dishId}")
     public ResponseEntity<?> deleteDish(@PathVariable Integer dishId) {
         try {
+            Dish dish = dishService.getDishById(dishId);
+            if (dish == null) {
+                return ResponseEntity.status(404).body("Dish not found with id " + dishId);
+            }
             dishService.deleteDish(dishId);
             return ResponseEntity.status(204).build();
         } catch (Exception e) {
             return ResponseEntity.status(500).body("Error deleting dish: " + e.getMessage());
+        }
+    }
+
+    @GetMapping("/restaurants/{id}/dishes")
+    public ResponseEntity<?> getDishesByRestaurantId(@PathVariable Integer id) {
+        try {
+            if (!restaurantService.existById(id)) {
+                return ResponseEntity.status(404).body("The restaurant with ID " + id + " does not exist.");
+            }
+            Restaurant restaurant = restaurantService.getRestaurantById(id);
+            return ResponseEntity.status(200).body(restaurant.getDishes());
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body("Error getting restaurant's dishes: " + e.getMessage());
+        }
+    }
+
+    @PatchMapping("/restaurants/{id}/dishes/{dishId}")
+    public ResponseEntity<?> updateDish(@PathVariable Integer id, @PathVariable Integer dishId, @RequestBody Map<String, Object> updates) {
+        try {
+//            Dish existingDish = dishService.getDishById(dishId);
+            Dish existingDish = dishService.getDishByIdAndRestaurantId(dishId,id);
+            if(existingDish == null){
+                return ResponseEntity.status(404).body("Dish " + dishId + " not found in restaurant " + id);
+            }
+
+            if (updates.containsKey("description")) {
+                existingDish.setDescription((String) updates.get("description"));
+            }
+            if (updates.containsKey("price")) {
+                existingDish.setPrice(((Integer) updates.get("price")).doubleValue());
+            }
+
+            dishService.updateDish(existingDish);
+            return ResponseEntity.status(200).body("Dish " + dishId + " updated successfully");
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body("Error updating dish: " + e.getMessage());
         }
     }
 }
